Two additions:  species, setPopupMode
added the notion of 'defaultSpecies' to CytoscapeWindow,
requiring a switch '-s' in CytoscapeConfig, and a companion
test in CytoscapeConfigTest

two new methods in CytoscapeWindow:

  public String getDefaultSpecies ()
  public String getSpecies (Node node)

for the first, the order of evaluation is

   1. if '-s' is on the command line, that value is returned
   2. if the props file has a 'species' value, that is returned
   3. 'unknown' is returned

for the second method, where a Node is the argument, there
is an additional rule evaluated first:

   0. if the node has an attribute named 'species', the value
      of that attribute is returned.

Most cytoscape analyses are done against a single species,
but at least in the case of node or edge annotation achieved
by inference via homology, multiple species may be represented
in the same graph.

Another new method:

  public void setPopupMode (PopupMode newMode)

This makes it easy to install new actions triggered
by a right-mouse clicked on the graph canvas.
I needed this to develop a new, more loosely wired
node browser; other plugins may need to customize
this behavior as well.