Introduced AbstractPlugin; here's the original email (2003.03.10)
I'd like to add AbstractPlugin.java to the core, as per
previous discussions of a plugin API. AbstractPlugin would be
the api to which plugins would need to subscribe; you can
check the draft version out in csplugins, under jarLoader/.
It basically has two methods: a one argument constructor,
taking a CytoscapeWindow, and a describe() function, as
suggested by Paul, I believe.

It's an abstract class rather than an interface because
we wanted to define the constructor in a certain way.
This means you have to call super() in your ctor, but it
doesn't do anything.  I thought of using reflection to
inspect the class for having a 1-argument CytoscapeWindow
constructor, but I concluded that I prefer having an
abstract class, even if it means one extra super() call.