Made changed to use the VisualStyle object to represent the current visual mappings as opposed to holding references to the node/edge appearance calculators.
Changed the vizmap loader of cytoscape to install a visual style
instead of node/edge appearance calculators.
Removed the node/edge appearance calculators from the CalculatorCatalog
and instead added VisualStyle objects to the catalog.
Changed the IO classes to work with VisualStyle objects.
Changed the VisualMappingManager to hold a VisualStyle and refer to it.
In the ui classes, changed the calls to get the node/edge appearance
calculator in accordance with the new VisualMappingManager interface.
The ui classes still hold and work with references to the node/edge
appearance calculators.
Updated the test class to be consistent with the other changes.

Also, preemtively fixed a potential bug by changing the
loadPropertyFileAsText method of CytoscapeConfig to send the text
through the load method of Properties instead of parsing the text
line-by-line.
-Andrew 03-20-03