Removal of most of the rest of the CytoscapeWindow references from the core. Rewrote the actions and dialogs to refer to CyWindow or NetworkView instead. With these references gone, was able to untangle the construction of CyMenus. Now the constructor of CMenus is called from an initialization method of CyWindow, and this constructor builds the basic menu bars without filling them with items. The same initialization method of CyWindow then immediately calls a method on CyMenus to fill the menus with items and action listeners. The old circular reference problem is now completely resolved.
Added constructors to CytoscapeWindow and CyWindow to allow
construction in either direction; i.e., one can first construct
CytoscapeWindow, when then builds a CyWindow, or first construct
the CyWindow which then creates a CytoscapeWindow wrapper. in
either case CytoscapeWindow's constructor triggers the plugin loading.

Added a method to GraphObjAttributes to import all data from
another GraphObjAttributes object; that is, it handles both
the attributes and the name mapping structures.
-Andrew 10-07-03