Fixed several bugs related to getting and setting the "size locked" flag from the ui.
On construction, this tab used to always set locked to true; now it
checks the state of the node appearance calculator.

When you switch to a different visual style, this object used to
not grab the new node appearance calculator, but kept working with
the old. I changed the visualStyleChanged() method to switch to
the new node appearance calculator, and also to check the locked
status on the new NAC and, if needed, resets the internal variable
and lock box selection state to be consistent.

Note that there is still a bug that, when you click the lock
check box, the UI doesn't change until you choose a different
tab and then go back.
-Andrew 10-13-2003