Moved a bunch of classes in the data/servers directory that were not being used to an archive directory. Rewrote the exclude tags of the compile target in build.xml to properly include the files that should be compiled (i.e., everything in data/servers and data/servers//unitTests) and exclude the proper ones (everything in the archive directories).
Added a constructor to CyNetwork and a method to CyNetworkFactory
to create a default network when no graph file is available.

Rewrote the cytoscape class and its constructor significantly. Removed
everything that wasn't being actively used, delegated to the factory
methods of CyNetworkFactory (allowing removal of all graph imports
from the cytoscape class) and changed the final call to construct
a CyWindow instead of a CytoscapeWindow.
-Andrew 11-04-03