Modified the network-related operations of AnnotationGui to notify before and after doing the operation.
For the one AnnotationGui operation that selects nodes in the graph,
added Giny support.

Changed the close dialog operation of AnnotationGui's dialog; now
the dialog stays around so you can see annotations that were
previously applied to the network.

Added a utility method to Semantics.

For the layout class, added a check in the doCallback method (the
one called by the AnnotationGUI) to do nothing and put up a message
dialog if the mode is Giny. I.e., the AttributeLayout class does
not support Giny graphs.
-Andrew 11-06-03