Update to plugin loading architecture. Various java doc comments updated in afffected files.
The purpose of these changes was to centralize the plugin loading
system so that plugins loaded in one window would be available in
other windows. Plugins are still attached to a CyWindow, but instead
of creating them directly there, plugins are instead loaded into
a registry (currently centralized in CytoscapeObj. This registry is
shared across an entire running instance of Cytoscape. Each window
will check its own plugin menu whenever the plugin menu is opened.
If there have been changes to the central registry since the last
time this window's menu has been updated, the menus is again updated
to include any plugins which have been loaded since the last update.
Missing plugins are instanciated with a link to the current CyWindow.

This plugin registry function will immediately be seperated out into
a new PluginRegistry class .. in an update to be made later today.

Other changes include:
- Dropping the support for conditional loading of plugins from the
cytoscape.props properties file. This allowed the code to be greatly
simplified, and that feature was not currently working correctly.
There are now 3 ways to load plugins: Through the user menu; with
the command line argument --JLD; and with an entry in the cytoscape.props
file of the form plugin.<name of plugin>.load<name of plugin>.

Changes were made to the startup scripts to allow the property file
loading method to work. Previously, these scripts used a command
like "java -jar cytoscape.jar" and this form of starting the VM
will *ignore* any CLASSPATH setting or -classpath command line
argument. Therefore, any class name which was included in the
properties file was not able to be found by the system classloader.
Instead, the startup scripts now do something like:
"java -classpath 'cytoscape.jar:.:$CLASSPATH' cytoscape.CyMain"
to start cytoscape. If CLASSPATH is set correctly, plugins can
now be loaded through the properties file.

Other functions have been consolidated in CyMenus. This is in
preparation for moving classes to a newly planned subpackage:
"cytoscape.plugin"