Added code to support flagging nodes and edges in a network plus test code. Modified CyNetwork and the implementation class to construct an instance of the flagger at construction and make it publicly accessible.
Added a class to synchronize the selection state in a view with the
flagged state in the displayed network, plus test code. Modified
the view implementation to use an instance of this class; also
reinsterted the GraphViewController into the view object so that
hidden objects are synchronized in the network and view.

Removed some hack code from the graphViewController that isn't
necessary now that the relevant method was added to Giny.

Modified the Cytoscape class to create a default CytoscapeObj and
to make the setCytoscapeObj method public.
-Andrew 2004-06-15