Code changes from Gregor Warsow. that allows Linkout on edges. Now we can use placeholders like %source.nodeAttribute%,%target.nodeAttribute%, %edgeAttribute%, %ID1%, %ID2%, %ID%.
line181: final String REGEX = "%.*%" was changed to final String REGEX = "%.*?%";
otherwise the pattern matching will be greedy and we are not able to
detect two or more placeholders to be substituted.

Lines274/275 was changed from:
    String fUrl = subsAttrs(url, attrs, sourceId, "ID%1", "source.");
to:
    String fUrl = subsAttrs(url, attrs, sourceId, "ID1", "source.");
The percent in square brackets in %ID[%]1% made it hard to detect the
right pattern.